src/main/java/ru/yandex/practikum/model/user/UserLogin.java
package ru.yandex.practikum.model.user;

public class UserLogin {
    private String email;
    private String password;

    public UserLogin(String email, String password) {
        this.email = email;
        this.password = password;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public static UserLogin getLogin(User user) {
        return new UserLogin(user.getEmail(), user.getPassword());
    }
}
